requirements.txt
torch==2.6.0
diffusers==0.35.1
transformers==4.41.0
accelerate>=0.20.0
streamlit==1.50.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow>=9.0.0
ftfy==6.2.0
einops==0.8.0